+++ model/user.go
@@ -4,7 +4,7 @@ import "gorm.io/gorm"
 
 type User struct {
 	gorm.Model
-	ID       int64
+	ID       uint
 	Email    string `gorm:"type:varchar(50); not null"`
 	Name     string `gorm:"type:varchar(20); not null"`
 	Password string `gorm:"type:varchar(100); not null"`
